requirements.txt
h5py==2.9.0
matplotlib==3.0.3
numpy==1.16.3
scipy
sympy==1.4
tensorflow-gpu==1.13.1